+++ processor/src/test/java/react4j/processor/AbstractReactProcessorTest.java
@@ -23,7 +23,7 @@ import static org.testng.Assert.*;
 @SuppressWarnings( "Duplicates" )
 abstract class AbstractReactProcessorTest
 {
-  void assertSuccessfulCompile( @Nonnull final String classname, final boolean needsHelper, final boolean dagger )
+  void assertSuccessfulCompile( @Nonnull final String classname, final boolean dagger )
     throws Exception
   {
     // It should be noted that we do not test the output of any Arez artifact
@@ -32,13 +32,14 @@ abstract class AbstractReactProcessorTest
     final StringBuilder input = new StringBuilder();
     final StringBuilder enhancedComponent = new StringBuilder();
     final StringBuilder builder = new StringBuilder();
-    final StringBuilder helper = new StringBuilder();
-    final StringBuilder daggerFactory = new StringBuilder();
+    final StringBuilder componentExtension = dagger ? new StringBuilder() : null;
     input.append( "input" );
     enhancedComponent.append( "expected" );
-    helper.append( "expected" );
     builder.append( "expected" );
-    daggerFactory.append( "expected" );
+    if ( dagger )
+    {
+      componentExtension.append( "expected" );
+    }
     for ( int i = 0; i < elements.length; i++ )
     {
       input.append( '/' );
@@ -49,37 +50,35 @@ abstract class AbstractReactProcessorTest
         enhancedComponent.append( "React4j_" );
       }
       enhancedComponent.append( elements[ i ] );
-      helper.append( '/' );
-      helper.append( elements[ i ] );
       builder.append( '/' );
       builder.append( elements[ i ] );
       if ( i == elements.length - 1 )
       {
-        helper.append( "_" );
         builder.append( "Builder" );
       }
-      daggerFactory.append( '/' );
-      daggerFactory.append( elements[ i ] );
-      if ( i == elements.length - 1 )
+      if ( dagger )
       {
-        daggerFactory.append( "DaggerFactory" );
+        componentExtension.append( '/' );
+        componentExtension.append( elements[ i ] );
+        if ( i == elements.length - 1 )
+        {
+          componentExtension.append( "DaggerComponentExtension" );
+        }
       }
     }
     input.append( ".java" );
     enhancedComponent.append( ".java" );
-    helper.append( ".java" );
     builder.append( ".java" );
-    daggerFactory.append( ".java" );
+    if ( dagger )
+    {
+      componentExtension.append( ".java" );
+    }
     final ArrayList<String> outputs = new ArrayList<>();
     outputs.add( enhancedComponent.toString() );
     outputs.add( builder.toString() );
-    if ( needsHelper )
-    {
-      outputs.add( helper.toString() );
-    }
     if ( dagger )
     {
-      outputs.add( daggerFactory.toString() );
+      outputs.add( componentExtension.toString() );
     }
     assertSuccessfulCompile( input.toString(), outputs.toArray( new String[ 0 ] ) );
   }