+++ CHANGELOG.md
@@ -97,6 +97,8 @@
   empty string will stop the GWT compiler from removing `toString()` on any subclass of `Object` (i.e. no
   `toString()` methods are candidates for dead code elimination).
 * **\[processor\]** Eliminate the static inner class named `Prop` from the enhanced react component if it is empty. 
+* **\[processor\]** Rename the generated dagger helper class from `[MyComponent]DaggerFactory` to
+  `[MyComponent]DaggerComponentExtension` to align with the same pattern in Arez.
 
 ### [v0.108](https://github.com/react4j/react4j/tree/v0.108) (2018-11-22)
 [Full Changelog](https://github.com/react4j/react4j/compare/v0.107...v0.108)